A variety of minor tweaks.
Fix paragraphs following the Trim: paragraph in the Complex Template
section to not look like sub-paragraphs of it.

Fix Curve Editing intro to not refer to "pressing a graph" in the Devo
10 manual.

Fix Transmitter power to indicate that it's possible to modify the 7E
to not get 7mW. Does a 7E with the diode mod still always display 150mW.

Fix model type to note that switching away from Helicopter instead of
to Plane switches to the Advanced GUI, as we now have a multi type.