Minor formatting/text fixes
Wording change in DSM2 protocol from "7 or less" to "7 or fewer".

Move the note about getting/where to find the license into the
paragraph about licensing.

Update the list of "other" protocols to the most popular, and note
that this is just a part of the list.

Remove Devo12 notes from Devo10/7E/12E manual.

Change warning about memory limitations to the Devo7, not Devo12.

Make bug reports a bulleted list, not inline (this is caused by a
missing blank line in front of the list, so presumably introduced when
converting to rst).